add unregister modal to kalypso dashboard
