+++ src/lib/atoms/componentClasses.ts
@@ -10,7 +10,7 @@ export const buttonClasses = {
 		'btn btn-theme btn-sm text-primary px-4 border-primary bg-inherit hover:text-primary-focus hover:border-primary-focus hover:bg-inherit disabled:bg-inherit disabled:text-grey-300 disabled:border-grey-300 disabled:cursor-not-allowed',
 	greyOutlined: `${baseFilledButtonClasses} border-[#D9DADE] bg-inherit hover:text-primary-focus hover:border-primary-focus hover:bg-inherit disabled:bg-inherit disabled:text-grey-300 disabled:border-grey-300 disabled:cursor-not-allowed`,
 	error:
-		'btn btn-theme btn-sm text-error px-5 border-error bg-inherit hover:bg-inherit hover:border-error hover:text-error-focus disabled:cursor-not-allowed',
+		'btn btn-theme btn-sm text-error px-5 border-error bg-inherit hover:bg-inherit hover:border-error hover:text-error-focus disabled:cursor-not-allowed disabled:bg-inherit disabled:text-error disabled:border-error disabled:opacity-60',
 	text: 'btn btn-sm btn-ghost p-0 bg-transparent hover:bg-transparent disabled:bg-transparent disabled:cursor-not-allowed',
 	icon: 'btn btn-xs btn-primary btn-circle p-0 rounded-full flex disabled:cursor-not-allowed disabled:bg-primary disabled:bg-opacity-60 disabled:text-white',
 	iconLightBlue:
